refactor(MetaTable): split Template.tsx into per-component modules [YTFRONT-5942]
