+++ packages/components/src/components/MetaTable/templates/TemplateReadable.tsx
@@ -1,0 +1,11 @@
+import {format as hammerFormat} from '../../../utils';
+
+import {itemBlock} from './utils';
+
+type Props = {
+    value?: string;
+};
+
+export function TemplateReadable({value = hammerFormat.NO_VALUE}: Props) {
+    return <span className={itemBlock('readable')}>{hammerFormat['ReadableField'](value)}</span>;
+}
