+++ packages/components/src/components/MetaTable/templates/index.ts
@@ -1,4 +1,7 @@
-export {TemplateId, TemplateFormattedValue, TemplateReadable, TemplateLink} from './Template';
+export {TemplateId} from './TemplateId';
+export {TemplateFormattedValue} from './TemplateFormattedValue';
+export {TemplateReadable} from './TemplateReadable';
+export {TemplateLink} from './TemplateLink';
 export {TemplateTime} from './TemplateTime';
 export type {
     TemplateTimeProps,
