+++ packages/components/src/components/MetaTable/templates/utils.ts
@@ -1,0 +1,3 @@
+import cn from 'bem-cn-lite';
+
+export const itemBlock: ReturnType<typeof cn> = cn('meta-table-item');
